I fix DRY.
