+++ hooks/useZoraCreateParameters.tsx
@@ -5,16 +5,17 @@ import { useAccount, usePublicClient } from 'wagmi'
 import { useSearchParams } from 'next/navigation'
 import getSalesConfig from '@/lib/zora/getSalesConfig'
 import useCreateMetadata from '@/hooks/useCreateMetadata'
+import useCreatorAddress from './useCreatorAddress'
 import { useProfileProvider } from '@/providers/ProfileProvider'
 
 const useZoraCreateParameters = (collection: Address) => {
   const publicClient = usePublicClient()
   const searchParams = useSearchParams()
   const { address } = useAccount()
-  const createMetadata = useCreateMetadata()
-  const payoutParam = searchParams.get('payoutRecipient')
-  const defaultAdmin = searchParams.get('defaultAdmin')
   const { profile } = useProfileProvider()
+  const createMetadata = useCreateMetadata()
+  const creatorAddress = useCreatorAddress()
+  const payoutParam = searchParams.get('payoutRecipient')
 
   const fetchParameters = async (chainId: number) => {
     if (!publicClient) return
@@ -23,9 +24,7 @@ const useZoraCreateParameters = (collection: Address) => {
     if (!cc0MusicIpfsHash) return
     const connnectedProfileAddress = profile?.connectedZoraProfile?.address
     const fallbackPayoutAddress = isAddress(payoutParam) ? payoutParam : address
-    const fallbackAccountAddress = isAddress(defaultAdmin) ? defaultAdmin : address
     const payoutRecipient = connnectedProfileAddress || fallbackPayoutAddress
-    const account = connnectedProfileAddress || fallbackAccountAddress
 
     const salesConfig = getSalesConfig(
       createMetadata.isTimedSale ? 'ZoraTimedSaleStrategy' : 'ZoraFixedPriceSaleStrategy',
@@ -41,7 +40,7 @@ const useZoraCreateParameters = (collection: Address) => {
           salesConfig,
           payoutRecipient,
         },
-        account,
+        account: creatorAddress,
       })
       newParameters = existingParameters
     } else {
@@ -56,7 +55,7 @@ const useZoraCreateParameters = (collection: Address) => {
           salesConfig,
           payoutRecipient,
         },
-        account,
+        account: creatorAddress,
       })
       newParameters = { ...newContractParameters, functionName: 'createContract' }
     }
